pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "clrmo-essentia-tagger"
version = "0.1.0"
description = "Analyze music files with Essentia and write tags/attributes back to a CLRMO database via the REST API"
requires-python = ">=3.9"
dependencies = [
    "essentia-tensorflow",
    "requests>=2.28",
]

[project.scripts]
clrmo-tag = "essentia_tagger.cli:main"
clrmo-download-models = "essentia_tagger.download_models:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["essentia_tagger*"]
